test 8b

+++ llama3_jax/simple_train.py
@@ -43,7 +43,7 @@ MODEL_NAME = "colab-llama-3.1-8B-Instruct-JAX"
 model_path, model, model_configurator, tokenizer = (
     automodel_lib.AutoJAXModelForCausalLM.from_pretrained(
         MODEL_NAME,
-        dtype=jnp.bfloat16,
+        dtype=float8,
         param_dtype=jnp.bfloat16,
         lora_rank=8,
         lora_alpha=16,
